Moved sparse and dense matricies to matrix
There's now a module that contains the submodules for both sparse and
dense matricies. This is to improve organization, especially since
before this commit the naming between sparse and dense matricies was
unclear. I'm not sure if I'm super happy with this change, or if sparse
and dense should both be in the library root.