Changed API so gmres takes in extra params
gmres now takes in the max number of iterations, the tolerance, and
the maximum number of search directions. There's an extra test to make
sure that we get an error if we exceed the max number of iterations, but
interestingly gmres appears to be very slow. As such the max iterations
is ~1000. This bears further investigation.

A bug was fixed in error.rs, I wasn't checking all of the cases right
for equality. There's also a new constructor for dense matricies that
takes rows instead of columns.