Iterator fixup
The all_elements iterator was very broken, and previously untested. It
has now been tested and appears to be working, and Display should be
working as a result. The implementation is somehow even hackier, and I
am unable to return a reference now. This a clone happens at each
iteration, which is really not optimal.

There's now a test to ensure that our iterator is actually working.